Reorganize Project (#9)
* Update requirements

* Reorganize into pubobot module

* Move updater to scripts/

* Use poetry to manage project

* Add ruff as a development dependency

* Auto-fix issues with ruff

* Remove trash, move docs

* Format entire project with ruff

* Allow files to be passed in via CLI

* Replace pyreadline with pyreadline3
