+++ README.md
@@ -1,1 +1,19 @@
-Nobody reads these
+# Pubobot
+
+## Development Setup
+
+[Install poetry for your system](https://python-poetry.org/docs/#installation).
+
+Create a `config.cfg` file and add the following:
+```python
+DISCORD_TOKEN = "your_discord_token"
+```
+
+For more options, see [config.cfg.example](config.cfg.example).
+
+Install dependencies and execute with poetry:
+
+```console
+$ poetry install
+$ poetry run pubobot
+```
