+++ doc/commands.md
