--- modules/__init__.py
@@ -1,1 +1,0 @@
-__all__ = ['irc', 'console', 'config', 'bot', 'stats2', 'scheduler']
