--- pugbot.bat
@@ -1,3 +1,0 @@
-cd C:\Test-PUGBOT\PugBot
-python pubobot.py
-pause
