--- requirements.txt
@@ -1,12 +1,0 @@
-aiohttp==3.6.2
-anyreadline==0.1.1
-async-timeout==3.0.1
-attrs==19.3.0
-chardet==3.0.4
-discord.py==1.5.0
-idna==2.8
-multidict==4.7.4
-pyreadline==2.1
-sqlite3-api==1.0.4
-websockets==8.1
-yarl==1.4.2
